requirements.txt
pytesseract==0.3.8
Pillow==9.3.0
gunicorn==20.1.0
Flask==2.1.0
werkzeug==2.1.0